Fix 18 exabyte memory allocation failure
The new_binary() function takes a size argument that is an
int. In the 64-bit emulator (sizeof(int) == 4, sizeof(Uint) == 8),
any sizes >= 0x8000000 become 0xffffffff80000000 and above and
triggers a memory allocation failure.

Change the type of the size argument to Uint, and change any
callers that cast the argument to an int.

Correction-by: Jon Meredith