Include fixture setup and cleanup errors in Eunit Surefire report
An error during fixture setup means that some tests could not be run,
and therefore needs to be highlighted in the test report.  Likewise, a
cleanup failure is often a problem that needs to be looked into.

Since setup and cleanup are not part of any single test in Eunit's
view, I include them as phantom test cases in the report whenever they
fail.