erts,hipe: Limited support for hipe cross compilation
Normally hipe assumes that generated native code will be executed
on the same emulator as the hipe compiler is running on.

This commit allows you to build a hipe "cross-compiler"
for a different configured emulator (but for same architecture).

Example:

> cd lib/hipe
> make clean
> make XCOMP=yes FLAVOR=smp TYPE=debug

This hipe compiler will generate code to be executed
on the debug compiled smp-emulator at $ERL_TOP.