v3_core: Fix variable incorrectly unbound after binary match
In the following code:

m(<<Sz:8,_:Sz/binary>>) ->
    Sz = wrong.

the Sz variable is supposed to be bound in the function header and the
matching "Sz = wrong" should cause a badarg exception.  But what
happens is that the Sz variables seems to be unbound and the matching
succeds and the m/1 function returns 'wrong'.

If the Sz variable is used directly (not matched), it will have
the expected value. Thus the following code:

m(<<Sz:8,_:Sz/binary>>) ->
    Sz.

will correctly return the value of Sz that was matched out from
the binary.

Reported-by: Bernard Duggan