Fix a bug concerning record field types
The default value 'undefined' was added to records field types in such
a way that the result was not always a well-formed type. This bug has
been fixed.

---

erl_pp has since OTP-8150 formatted types so that 'undefined' was
removed from union types assigned to record fields. Since one cannot
distinguish between 'undefined' added by the parser or supplied by the
user, a side effect was that user supplied 'undefined's were also
removed.

Now the pretty printer shows 'undefined' even if added by the parser.
This is a minor issue.