Never fail when stopping rb, and fix file descriptor leak
rb:stop did sometimes return {error,running}.
This came from supervisor:delete_child and happened when the rb_server
has not yet terminated when this function was called. Instead of
having a separate gen_server call to rb_server for stopping the
process, supervisor:terminate_child is now called. This is a
synchronous function - i.e. it waits for the process to actually
terminate before it returns.

A file descriptor leak in rb:scan_files is corrected. The index file
was never closed after reading.