Completed bug fix "temporary child specs should not be kept when child terminates" and improved test suite
The bug fix supplied by Filipe David Manana <fdmanana@apache.org>
did not cover all possible ways that a process may be terminated
as for instance with supervisor:terminate_child. Also there
was a bug in the base case of the patch returning a list of a list instead
of only the list.

Added a timeout for the test cases, eliminated unnecessary sleeps,
improved code.