Get rid of redundant 'try_me_else' and 'fail' instructions
If the left part of a transformation will always match, omit the
the 'try_me_else' and 'fail' instructions.

As part of this optimization, make it an error to have a
transformation that can never be reached because of a previous
transformation that will always match. (Remove one transformation
from ops.tab that was found to be unreachable.)