erts: Fix dependency generation
Apparantly, an invocation like "make generate depend" will run the
'generate' and 'depend' targets in parallel, even if the Makefile
contains a .NOTPARALLEL target.  In this case, it will result in
incorrect dependencies since not all files have been generated when
the dependencies are computed.

Split up the single 'make' invocation to two invocations, to force the
targets to be run sequentially.