Do not add an empty radio box on the releases tab for the start_clean release
First, the radiobox is changed to a listbox, since this will allow
multiple selections. This is however for future use - for now a
selection will only cause a printout in the erlang shell.

Second, add kernel and stdlib to the list of applications in order to
make the picture complete and avoid an empty list (radio) box for the
start_clean release.