Let epmd ignore empty ERL_EPMD_ADDRESS
If the environment variable ERL_EPMD_ADDRESS is set to the empty string,
empd now behaves like it does by default when ERL_EPMD_ADDRESS is unset.
That is, in this case, epmd now listens on all available interfaces
instead of using only the loopback interface, which happened because
epmd added the loopback address to the (in this case empty) list of
addresses specified via ERL_EPMD_ADDRESS.

Also, epmd now ignores ERL_EPMD_ADDRESS if it contains only separator
characters (comma and space).

The same applies to epmd's -address option.