Fix enif_compare on 64bits machines
In 64bits machines the Sint type has a size of 8 bytes,
while on 32bits machines it has a 4 bytes size.
enif_compare was ignoring this and therefore returning
incorrect values when the result of the CMP function
(which returns a Sint value) doesn't fit in 4 bytes.

For example, passing the operands -1294536544000 and
-1178704800000 to enif_compare would trigger the bug.