Add option purge to release_handler:check_install_release
Given this option, all modules that are to be purged by indicated
upgrade,and that can be soft purged, will be purged when all other
check of check_install_release have been successfully completed.

I added a note under install_release in the reference manual about how
to use check_install_release with this new option in order to speed up
the execution of install_release.

I also added three more test cases for this functionality.