Explicitly kill dynamic children in supervisors
According to the supervisor's documentation:
  "Important note on simple-one-for-one supervisors: The dynamically
   created child processes of a simple-one-for-one supervisor are not
   explicitly killed, regardless of shutdown strategy, but are expected
   to terminate when the supervisor does (that is, when an exit signal
   from the parent process is received)."

All is fine as long as we stop simple_one_for_one supervisor manually.
Dynamic children catch the exit signal from the supervisor and leave.
But, if this happens when we stop an application, after the top
supervisor has stopped, the application master kills all remaining
processes associated to this application. So, dynamic children that trap
exit signals can be killed during their cleanup (here we mean inside
terminate/2). This is unpredictable and highly time-dependent.

In this commit, supervisor module is patched to explicitly terminate
dynamic children accordingly to the shutdown strategy.

NOTE: Order in which dynamic children are stopped is not defined. In
fact, this is "almost" done at the same time.